Create directory modules with  add, remove, render, storage functions. Add ToDo Model and call all functions as methods. Create a new object. Change span element for description to textarea.
