+++ src/styles.css
@@ -50,10 +50,6 @@ body {
   transition: transform 0.5s ease;
 }
 
-.container:hover {
-  transform: translateY(-10px);
-}
-
 .tasks {
   display: flex;
   flex-direction: column;
@@ -177,4 +173,16 @@ li input[type=checkbox]:checked + span {
   outline: none;
   border: none;
   margin: 1rem auto;
+}
+
+textarea {
+  width: 100%;
+  resize: none;
+  outline: none;
+  border: none;
+}
+
+.completed {
+  text-decoration: line-through;
+  color: #999;
 }
